requirements.txt
tweepy
ffmpeg-python
Pillow
configparser
Flask
Flask-RESTful